wafec_tests_openstack_testware/requirements.txt
python-novaclient>=17.2.1
python-glanceclient>=3.2.2
python-neutronclient>=7.2.1
python-keystoneclient>=4.1.1
Flask>=1.1.2
SQLAlchemy>=1.3.20
PyMySQL>=0.10.1